third_party/WebKit/LayoutTests/svg/dom/SVGTransformTearOff-contextElement-crash.html

<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

var svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
var rect = document.createElementNS("http://www.w3.org/2000/svg", "rect");
rect.transform.baseVal.appendItem(svg.createSVGTransform());
var rectObservation = internals.observeGC(rect);
var transform = rect.transform.baseVal.getItem(0);
rect = null;
gc();

transform.angle; // This should not cause crash in ASAN.

// rect should be kept alive from "transform"
shouldBeFalse('rectObservation.wasCollected');
</script>
This test passes if it doesn't crash in ASAN.
</body>
</html>
